+++ package.json
@@ -7,9 +7,11 @@
   "dependencies": {
     "booru": "^0.2.0",
     "cheerio": "^0.22.0",
+    "common-tags": "^1.4.0",
     "countdown": "^2.6.0",
     "cydia-api-node": "^1.0.2",
     "discord.js": "^11.1.0",
+    "discord.js-commando": "^0.9.0",
     "malapi": "0.0.3",
     "moment": "^2.17.1",
     "omdb": "^0.8.0",
@@ -17,6 +19,7 @@
     "querystring": "^0.2.0",
     "request": "^2.79.0",
     "scalc": "^2.0.1",
+    "sqlite": "^2.7.0",
     "superagent": "^3.4.4",
     "urban": "^0.3.1",
     "youtube-node": "^1.3.0"
