--- pyrrharefresh.sh
@@ -1,3 +1,0 @@
-pm2 restart 2
-pm2 reset 2
-pm2 flush
