+++ src/commands/reactions/images/opinion.gif
